Remove Commands.run
